pyproject.toml
[tool.poetry]
name = "multistore"
version = "0.1.0"
description = ""
authors = ["Bakdaulet"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
django-filter = "^22.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.1"
portodjangostarter = "^0.1.4"
gunicorn = "^20.1.0"
requests = "^2.28.1"
Pillow = "^9.2.0"
django-cors-headers = "^3.13.0"
loguru = "^0.6.0"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^1.7.0"


[tool.poe.poetry_hooks]
serve = "python manage.py runserver"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
